requirements.txt
atomicwrites==1.4.1 ; sys_platform == 'win32'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
boto3==1.21.46
botocore==1.24.46 ; python_version >= '3.6'
cachetools==5.3.1 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
colorama==0.4.6 ; sys_platform == 'win32'
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-stripe==2.5.1
django==2.2.28
django-allauth==0.50.0
django-bootstrap4==22.1
django-cleanup==8.0.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.3
factory-boy==3.2.1
faker==18.13.0 ; python_version >= '3.7'
google-api-core==2.10.2 ; python_version >= '3.7'
google-auth==2.6.2
google-cloud-core==2.3.3 ; python_version >= '3.7'
google-cloud-secret-manager==2.9.2
google-cloud-storage==2.2.1
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.59.1 ; python_version >= '3.7'
googlemaps==4.10.0
grpc-google-iam-v1==0.12.6 ; python_version >= '3.7'
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
jsonfield==3.1.0 ; python_version >= '3.6'
markupsafe==2.1.3 ; python_version >= '3.7'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==23.1
pillow==9.1.1
pluggy==1.2.0 ; python_version >= '3.7'
proto-plus==1.22.3 ; python_version >= '3.6'
protobuf==4.23.4 ; python_version >= '3.7'
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pyotp==2.8.0
pytest==7.1.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2023.3
pyyaml==6.0
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
ruamel.yaml==0.17.32 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.12' and platform_python_implementation == 'CPython'
s3transfer==0.5.2 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
stripe==5.4.0
tomli==2.0.1 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
waitress==2.1.2
whitenoise==6.0.0
psycopg2-binary>=2.8,<2.9
django-smtp-ssl==1.0
